+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `EventDto.serialize` function
 
 ## 10.0.0 - 2010-05-25
 - [**BC**] Use .netcore 3.1